refactor(models): move matrix cockpit scanner and stock contracts
