+++ src/uw_scan/models/__init__.py
@@ -39,6 +39,44 @@ from .options import (
     OptionContractRow,
     OptionsDailyRow,
 )
+from .cockpit import (
+    CharmSignal,
+    CockpitDealerMetrics,
+    CockpitDealerPoint,
+    CockpitDealerResponse,
+    CockpitFlowAlert,
+    CockpitFlowImResponse,
+    CockpitImPoint,
+    CockpitSkewPoint,
+    CockpitStateResponse,
+    CockpitSurfaceResponse,
+    CockpitTermPoint,
+    CockpitVrpPoint,
+    CockpitVrpResponse,
+    VannaSignal,
+)
+from .matrix import MatrixSourceFreshness, MatrixState, SetupClassification
+from .scanner import (
+    BulkScreenerRow,
+    EtfInfo,
+    EtfInOutflowRow,
+    GexLevel,
+    MarketAggregates,
+    MarketStructureLevels,
+    ScanReport,
+    ScanTickerResult,
+    StrikeGexBucket,
+)
+from .stock import (
+    MarketStructure,
+    SingleStockReport,
+    StockHistoryResponse,
+    StockHistoryRow,
+    TradePlan,
+    TradePlanLeg,
+    VRPAssessment,
+    VolatilityProfile,
+)
 
 
 # ---------------------------------------------------------------------------
@@ -126,521 +164,11 @@ class SkewRow(_UwBase):
 # ---------------------------------------------------------------------------
 # Aggregates for the Single-Stock Card
 # ---------------------------------------------------------------------------
-class MarketStructure(_UwBase):
-    spot: Decimal | None = None
-    nearest_expiry: _date | None = None
-    total_call_gex: Decimal | None = None
-    total_put_gex: Decimal | None = None
-    net_gex: Decimal | None = None
-    total_call_dex_oi: Decimal | None = None
-    total_put_dex_oi: Decimal | None = None
-    max_pain: Decimal | None = None
-    top_call_oi_strikes: list[Decimal] = []
-    top_put_oi_strikes: list[Decimal] = []
-
-
-class VolatilityProfile(_UwBase):
-    iv: Decimal | None = None
-    iv_rank: Decimal | None = None
-    iv_low_52w: Decimal | None = None
-    iv_high_52w: Decimal | None = None
-    rv: Decimal | None = None
-    rv_low_52w: Decimal | None = None
-    rv_high_52w: Decimal | None = None
-    iv_rank_1y: Decimal | None = None
-    iv_percentile_30d: Decimal | None = None
-    implied_move_30d_perc: Decimal | None = None
-    skew_25d: Decimal | None = None
-    term_dte_to_iv: list[tuple[int, Decimal]] = []
-
-
-class VRPAssessment(_UwBase):
-    vrp: Decimal | None = None
-    signal: str
-    note: str
-
-
-class TradePlanLeg(_UwBase):
-    option_symbol: str
-    side: str  # "buy" / "sell"
-    strike: Decimal
-    expiry: _date
-    mid: Decimal | None = None
-
-
-class TradePlan(_UwBase):
-    structure: str
-    direction: str
-    legs: list[TradePlanLeg] = []
-    rationale: str
-    max_loss: Decimal | None = None
-    max_profit: Decimal | None = None
-
-
-class SetupClassification(_UwBase):
-    setup_type: str  # "C"
-    label: str  # "Deep Conviction"
-    direction: str  # "bull" / "bear"
-    score: Decimal
-    confirmations: list[str] = []
-    warnings: list[str] = []
-    notes: str = ""
-
-
-class MatrixState(_UwBase):
-    ticker: str
-    market_date: _date
-    threshold_version: int = 1
-    vanna_state: MatrixDirection
-    charm_state: MatrixDirection
-    skew_state: MatrixDirection
-    term_state: MatrixDirection
-    im_state: MatrixDirection
-    flow_state: MatrixDirection
-    vrp_state: MatrixDirection
-    consistency_tier: MatrixConsistencyTier
-    cluster_coverage_ok: bool
-    term_classification: (
-        Literal["contango", "event_back", "liquidity_back", "mixed"] | None
-    ) = None
-    skew_25d_zscore_180d: Decimal | None = None
-    iv_atm_30d: Decimal | None = None
-    rv_30d: Decimal | None = None
-    vrp: Decimal | None = None
-    vrp_zscore_60d: Decimal | None = None
-    implied_move_pct: Decimal | None = None
-    front_iv: Decimal | None = None
-    back_iv: Decimal | None = None
-    front_back_spread: Decimal | None = None
-    pin_distance_sigma: Decimal | None = None
-    vrp_sign_flip_status: bool | Literal["insufficient_history"] = (
-        "insufficient_history"
-    )
-    vrp_sign_flip_aligned_days: int = 0
-    vanna_conditional_reading: VannaConditionalReading | None = None
-    directional_imbalance_3d: Decimal | None = None
-    vanna_oi_change_bias: Literal["call_oi_build", "put_oi_build", "mixed"] | None = (
-        None
-    )
-    charm_regime: CharmRegime | None = None
-    charm_stress_override: bool = False
-    skew_25d_5d_change: Decimal | None = None
-    skew_regime: SkewRegime | None = None
-    skew_term_structure: Decimal | None = None
-    single_point_bump_pct: Decimal | None = None
-    full_curve_slope_pct: Decimal | None = None
-    term_johnson_slope_pc1: Decimal | None = None
-    atm_straddle_mid: Decimal | None = None
-    implied_move_expected_abs: Decimal | None = None
-    implied_move_event_percentile: Decimal | None = None
-    vrp_zscore_252d: Decimal | None = None
-
-
-class MatrixSourceFreshness(_UwBase):
-    vanna_charm: datetime | None = None
-    skew: datetime | None = None
-    term: datetime | None = None
-    im_vrp: datetime | None = None
-    vrp_rv: datetime | None = None
-    oi: datetime | None = None
-
-
-class CockpitStateResponse(_UwBase):
-    state: MatrixState
-    freshness: MatrixSourceFreshness
-
-
-class CockpitDealerPoint(_UwBase):
-    expiry: _date
-    strike: Decimal
-    call_vanna: Decimal | None = None
-    put_vanna: Decimal | None = None
-    call_charm: Decimal | None = None
-    put_charm: Decimal | None = None
-    exposure_call_vanna: Decimal | None = None
-    exposure_put_vanna: Decimal | None = None
-    exposure_call_charm: Decimal | None = None
-    exposure_put_charm: Decimal | None = None
-
-
-class CockpitDealerMetrics(_UwBase):
-    pin_candidate_strike: Decimal | None = None
-    pin_candidate_expiry: _date | None = None
-    pin_source_date: _date | None = None
-    pin_distance_sigma: Decimal | None = None
-    pin_regime_flag: bool | None = None
-    dealer_net_vanna_proxy: Decimal | None = None
-    dealer_net_charm_proxy: Decimal | None = None
-    flow_color_lookback_3d: Literal["put_heavy", "call_heavy", "neutral"] | None = None
-    flow_put_premium_3d: Decimal | None = None
-    flow_call_premium_3d: Decimal | None = None
-    iv_30d_delta_5d: Decimal | None = None
-    net_gamma: Decimal | None = None
-    net_gamma_sign: Literal["positive", "negative", "neutral"] | None = None
-    gamma_regime: Literal["long_gamma", "short_gamma", "neutral"] | None = None
-    vanna_conditional_reading: VannaConditionalReading | None = None
-    directional_imbalance_3d: Decimal | None = None
-    vanna_oi_change_bias: Literal["call_oi_build", "put_oi_build", "mixed"] | None = (
-        None
-    )
-    charm_regime: CharmRegime | None = None
-    charm_stress_override: bool | None = None
-
-
-class VannaSignal(_UwBase):
-    ticker: str
-    market_date: _date
-    dealer_net_vanna_proxy: Decimal | None = None
-    flow_color_lookback_3d: Literal["put_heavy", "call_heavy", "neutral"] | None = None
-    flow_put_premium_3d: Decimal | None = None
-    flow_call_premium_3d: Decimal | None = None
-    iv_30d_delta_5d: Decimal | None = None
-    vanna_conditional_reading: VannaConditionalReading | None = None
-    directional_imbalance_3d: Decimal | None = None
-    vanna_oi_change_bias: Literal["call_oi_build", "put_oi_build", "mixed"] | None = (
-        None
-    )
-    generated_at: datetime | None = None
-    inserted_at: datetime | None = None
-
-
-class CharmSignal(_UwBase):
-    ticker: str
-    market_date: _date
-    pin_candidate_strike: Decimal | None = None
-    pin_candidate_expiry: _date | None = None
-    pin_source_date: _date | None = None
-    pin_distance_sigma: Decimal | None = None
-    pin_regime_flag: bool | None = None
-    dealer_net_charm_proxy: Decimal | None = None
-    net_gamma: Decimal | None = None
-    net_gamma_sign: Literal["positive", "negative", "neutral"] | None = None
-    gamma_regime: Literal["long_gamma", "short_gamma", "neutral"] | None = None
-    charm_regime: CharmRegime | None = None
-    charm_stress_override: bool | None = None
-    generated_at: datetime | None = None
-    inserted_at: datetime | None = None
-
-
-class CockpitDealerResponse(_UwBase):
-    ticker: str
-    market_date: _date
-    metrics: CockpitDealerMetrics = Field(default_factory=CockpitDealerMetrics)
-    points: list[CockpitDealerPoint] = Field(default_factory=list)
-
-
-class CockpitSkewPoint(_UwBase):
-    market_date: _date
-    expiry: _date | None = None
-    risk_reversal: Decimal | None = None
-
-
-class CockpitTermPoint(_UwBase):
-    expiry: _date
-    dte: int | None = None
-    volatility: Decimal | None = None
-    implied_move_perc: Decimal | None = None
-    implied_move_expected_abs: Decimal | None = None
-
-
-class CockpitSurfaceResponse(_UwBase):
-    ticker: str
-    market_date: _date
-    skew: list[CockpitSkewPoint] = Field(default_factory=list)
-    term: list[CockpitTermPoint] = Field(default_factory=list)
-
-
-class CockpitFlowAlert(_UwBase):
-    alert_id: str
-    option_chain: str | None = None
-    expiry: _date | None = None
-    strike: Decimal | None = None
-    option_type: str | None = None
-    total_premium: Decimal | None = None
-    volume: int | None = None
-    open_interest: int | None = None
-    total_ask_side_prem: Decimal | None = None
-    total_bid_side_prem: Decimal | None = None
-    has_sweep: bool | None = None
-    has_floor: bool | None = None
-    has_multileg: bool | None = None
-    all_opening_trades: bool | None = None
-    alert_rule: str | None = None
-    flow_footprint_label: FlowFootprintLabel | None = None
-    aggressor_label_confidence: Decimal | None = None
-    created_at: datetime | None = None
-
-
-class CockpitImPoint(_UwBase):
-    market_date: _date
-    days: int
-    volatility: Decimal | None = None
-    implied_move_perc: Decimal | None = None
-    implied_move_expected_abs: Decimal | None = None
-    percentile: Decimal | None = None
-
-
-class CockpitFlowImResponse(_UwBase):
-    ticker: str
-    market_date: _date
-    alerts: list[CockpitFlowAlert] = Field(default_factory=list)
-    implied_moves: list[CockpitImPoint] = Field(default_factory=list)
-
-
-class CockpitVrpPoint(_UwBase):
-    market_date: _date
-    iv: Decimal | None = None
-    rv: Decimal | None = None
-    vrp: Decimal | None = None
-    iv_rank_1y: Decimal | None = None
-
-
-class CockpitVrpResponse(_UwBase):
-    ticker: str
-    market_date: _date
-    points: list[CockpitVrpPoint] = Field(default_factory=list)
 
 
 # ---------------------------------------------------------------------------
 # Bulk screener row (S2) — `/api/screener/stocks`
 # ---------------------------------------------------------------------------
-class BulkScreenerRow(_UwBase):
-    ticker: str
-    date: _date | None = None
-    sector: str | None = None
-    issue_type: str | None = None
-    full_name: str | None = None
-    is_index: bool | None = None
-    er_time: str | None = None
-    next_earnings_date: _date | None = None
-    next_dividend_date: _date | None = None
-    marketcap: Decimal | None = None
-    # Prices / volume
-    close: Decimal | None = None
-    prev_close: Decimal | None = None
-    high: Decimal | None = None
-    low: Decimal | None = None
-    week_52_high: Decimal | None = None
-    week_52_low: Decimal | None = None
-    stock_volume: int | None = None
-    avg30_volume: Decimal | None = None
-    relative_volume: Decimal | None = None
-    # Premiums
-    call_premium: Decimal | None = None
-    put_premium: Decimal | None = None
-    net_call_premium: Decimal | None = None
-    net_put_premium: Decimal | None = None
-    bullish_premium: Decimal | None = None
-    bearish_premium: Decimal | None = None
-    # Flow context
-    put_call_ratio: Decimal | None = None
-    call_volume: int | None = None
-    put_volume: int | None = None
-    call_volume_ask_side: int | None = None
-    call_volume_bid_side: int | None = None
-    put_volume_ask_side: int | None = None
-    put_volume_bid_side: int | None = None
-    # OI
-    call_open_interest: int | None = None
-    put_open_interest: int | None = None
-    total_open_interest: int | None = None
-    prev_call_oi: int | None = None
-    prev_put_oi: int | None = None
-    # IV / vol
-    iv_rank: Decimal | None = None
-    iv30d: Decimal | None = None
-    iv30d_1d: Decimal | None = None
-    iv30d_1w: Decimal | None = None
-    iv30d_1m: Decimal | None = None
-    volatility: Decimal | None = None
-    volatility_7: Decimal | None = None
-    volatility_30: Decimal | None = None
-    realized_volatility: Decimal | None = None
-    variance_risk_premium: Decimal | None = None
-    # Implied move
-    implied_move: Decimal | None = None
-    implied_move_7: Decimal | None = None
-    implied_move_30: Decimal | None = None
-    implied_move_perc: Decimal | None = None
-    implied_move_perc_7: Decimal | None = None
-    implied_move_perc_30: Decimal | None = None
-    # GEX
-    gex_net_change: Decimal | None = None
-    gex_ratio: Decimal | None = None
-    gex_perc_change: Decimal | None = None
-    cum_dir_delta: int | None = None
-    cum_dir_gamma: int | None = None
-    cum_dir_vega: int | None = None
-
-
-class EtfInfo(_UwBase):
-    aum: Decimal | None = None
-    name: str | None = None
-    avg30_volume: Decimal | None = None
-    has_options: bool | None = None
-
-
-class EtfInOutflowRow(_UwBase):
-    ticker: str
-    date: _date
-    change: Decimal | None = None
-    change_prem: Decimal | None = None
-    close: Decimal | None = None
-    volume: Decimal | None = None
-    expiration_cycle: str | None = None
-    is_fomc: bool | None = None
-
-
-class ScanTickerResult(_UwBase):
-    """One row in the Full Scan output. Ranked by `score` desc."""
-
-    ticker: str
-    setup_type: str | None = None  # "C", "F", or None
-    label: str | None = None
-    direction: str | None = None
-    score: Decimal = Decimal("0")
-    net_premium: Decimal | None = None
-    net_call_premium: Decimal | None = None
-    net_put_premium: Decimal | None = None
-    iv_rank: Decimal | None = None
-    sector: str | None = None
-    relative_volume: Decimal | None = None
-    gex_net_change: Decimal | None = None
-    variance_risk_premium: Decimal | None = None
-    total_open_interest: int | None = None
-    next_earnings_date: _date | None = None
-    signals_present: list[str] = []
-    confirmations: list[str] = []
-    warnings: list[str] = []
-    notes: str = ""
-    screener_row: BulkScreenerRow | None = None
-
-
-class ScanReport(_UwBase):
-    run_id: int
-    generated_at: datetime
-    scan_date: _date | None = None
-    universe_size: int
-    universe_returned: int
-    results: list[ScanTickerResult] = []
-    dropped_tickers: list[str] = []
-    top_pick: str | None = None
-
-
-class MarketAggregates(_UwBase):
-    """Per-ticker aggregate fields sourced from the bulk-screener endpoint.
-
-    Populated by pipeline.run_single_stock alongside the existing per-section
-    sub-models. Feeds the watchlist card POSITIONING and SKEW blocks.
-    """
-
-    call_oi_total: int | None = None
-    put_oi_total: int | None = None
-    call_volume_total: int | None = None
-    put_volume_total: int | None = None
-    call_volume_ask_side: int | None = None
-    call_volume_bid_side: int | None = None
-    put_volume_ask_side: int | None = None
-    put_volume_bid_side: int | None = None
-    pcr_oi: Decimal | None = None
-    pcr_vol: Decimal | None = None
-    iv30d: Decimal | None = None
-    market_cap: Decimal | None = None
-    aum: Decimal | None = None
-
-
-class StrikeGexBucket(_UwBase):
-    """One row of the per-strike, per-expiry GEX curve persisted on each scan run."""
-
-    strike: Decimal
-    expiry: _date
-    net_gex: Decimal | None = None
-    call_gex: Decimal | None = None
-    put_gex: Decimal | None = None
-
-
-class GexLevel(_UwBase):
-    """One labeled level on the GEX curve (e.g. CALL WALL, PUT WALL, MAX MAGNET).
-
-    `gamma_per_dollar` is the per-strike net_gex used as the "$N per $1" sensitivity
-    figure on the tile — the dollar value of dealer hedging triggered by a $1 move.
-    """
-
-    strike: Decimal
-    net_gex: Decimal | None = None
-    pct_from_spot: Decimal | None = None
-    gamma_per_dollar: Decimal | None = None
-
-
-class MarketStructureLevels(_UwBase):
-    """Derived strike-level reference points used by the Market Structure tab.
-
-    Conventions follow FlashAlpha / SpotGamma:
-      - gex_flip: lowest strike where running cumulative net_gex flips sign
-      - call_wall: strike with largest call-side gamma (typically above spot)
-      - put_wall: strike with largest put-side gamma magnitude (typically below spot)
-      - max_magnet: strike with largest positive net_gex above spot (pulls price up)
-      - second_magnet: strike with second-largest positive net_gex above spot
-      - max_accel: strike with most-negative net_gex below the flip (movement accelerator)
-    """
-
-    gex_flip: GexLevel | None = None
-    call_wall: GexLevel | None = None
-    put_wall: GexLevel | None = None
-    max_magnet: GexLevel | None = None
-    second_magnet: GexLevel | None = None
-    max_accel: GexLevel | None = None
-
-
-class StockHistoryRow(_UwBase):
-    """One per-trading-day rollup of a ticker's market structure.
-
-    Built from the latest successful scan_run on that date. spot comes from
-    daily_ohlc.close so it's a stable end-of-day reference (NULL for the
-    current trading day until the OHLC pull fires post-close).
-    """
-
-    market_date: _date
-    spot: Decimal | None = None
-    gex_flip: Decimal | None = None
-    net_gex: Decimal | None = None
-    net_dex: Decimal | None = None
-    iv30d: Decimal | None = None
-    pcr_vol: Decimal | None = None
-    bias: str = "NEUTRAL"
-
-
-class StockHistoryResponse(_UwBase):
-    ticker: str
-    rows: list[StockHistoryRow] = []
-
-
-class SingleStockReport(_UwBase):
-    run_id: int
-    ticker: str
-    generated_at: datetime
-    spot_quoted_at: datetime | None = None
-    spot_source: str | None = None
-    short_int_note: str = "n/a (UW endpoint does not expose %)"
-    market_structure: MarketStructure
-    volatility: VolatilityProfile
-    flow: FlowSnapshot
-    vrp: VRPAssessment
-    setup: SetupClassification | None = None
-    trade_plan: TradePlan | None = None
-    dark_pool_notional: Decimal | None = None
-    dark_pool_print_count: int = 0
-    short_data: ShortDataRow | None = None
-    max_pain_rows: list[MaxPainRow] = []
-    oi_change_top: list[OiChangeRow] = []
-    aggregates: MarketAggregates | None = None
-    strike_gex_curve: list[StrikeGexBucket] = []
-    market_structure_levels: MarketStructureLevels | None = None
-    options_timeline: list[OptionsDailyRow] = []
-    option_chain_per_strike: list[OptionChainPerStrikeRow] = []
-    # Promoted from FlowAlert.next_earnings_date so the Volume-timeline panel
-    # can render the earnings marker without iterating alerts on the client.
-    next_earnings_date: _date | None = None
 
 
 # ---------------------------------------------------------------------------
